linux: keep update_pages() lengths wide enough for clone repair
update_pages() repairs the mapped file view after operations such as
clone-range. Its length parameter was still typed as int even though
callers and offsets are 64-bit.

Widen the helper to uint64_t on both Linux and FreeBSD so large repair
ranges cannot be truncated before the page update walk runs.

+++ module/os/freebsd/zfs/zfs_vnops_os.c
@@ -393,7 +393,7 @@ page_unhold(vm_page_t pp)
  *		the page and the dmu buffer.
  */
 void
-update_pages(znode_t *zp, int64_t start, int len, objset_t *os)
+update_pages(znode_t *zp, int64_t start, uint64_t len, objset_t *os)
 {
 	vm_object_t obj;
 	struct sf_buf *sf;
